requirements.txt
asgiref==3.11.0
attrs==25.4.0
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
contourpy==1.3.3
cryptography==46.0.3
cycler==0.12.1
Django==5.2.8
django-cors-headers==4.9.0
django-environ==0.12.0
django-filter==25.2
django-oauth-toolkit==3.1.0
djangorestframework==3.16.1
drf-spectacular==0.29.0
et_xmlfile==2.0.0
fonttools==4.61.0
gunicorn==23.0.0
idna==3.11
inflection==0.5.1
joblib==1.5.2
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jwcrypto==1.5.6
kiwisolver==1.4.9
matplotlib==3.10.7
mysqlclient==2.2.7
narwhals==2.13.0
numpy==2.3.5
oauthlib==3.3.1
openpyxl==3.1.5
packaging==25.0
pandas==2.3.3
patsy==1.0.2
pillow==12.0.0
plotly==6.5.0
pycparser==2.23
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
PyYAML==6.0.3
referencing==0.37.0
requests==2.32.5
rpds-py==0.30.0
scikit-learn==1.7.2
scipy==1.16.3
seaborn==0.13.2
six==1.17.0
sqlparse==0.5.3
statsmodels==0.14.5
threadpoolctl==3.6.0
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
whitenoise==6.11.0
